hive sdk changes and refactor
