+++ metastor/managedObjectLoader/pom.xml
@@ -19,7 +19,7 @@
 	<parent>
 		<artifactId>metastore</artifactId>
 		<groupId>org.finra.herd-mdl.metastore</groupId>
-		<version>1.2.51</version>
+		<version>1.2.52</version>
 	</parent>
 
 	<modelVersion>4.0.0</modelVersion>
