+++ metastor/managedObjectLoader/src/main/java/org/finra/herd/metastore/managed/jobProcessor/HiveHqlGenerator.java
@@ -67,6 +67,12 @@ public class HiveHqlGenerator {
     @Autowired
     JobProcessorDAO jobProcessorDAO;
 
+    @Autowired
+    DetectSchemaChanges detectSchemaChanges;
+
+    @Autowired
+    HiveAlterTable hiveAlterTable;
+
 
     public List<String> schemaSql(boolean schemaExists, JobDefinition jd) throws ApiException, SQLException {
 
@@ -89,10 +95,10 @@ public class HiveHqlGenerator {
 
                 try {
 
-                    FormatChange change = detectSchemaChange(jd, hiveTableSchema, format, ddl);
-                    formatRegularColumn(change,jd,list,tableName);
-                    formatPartitionColumn(change,jd,list,tableName);
-                    formatClusterColumn(change,jd,list,tableName);
+                    FormatChange change = detectSchemaChanges.detectSchemaChange(jd, hiveTableSchema, format, ddl);
+                    hiveAlterTable.formatRegularColumn(change,jd,list,tableName);
+                    hiveAlterTable.formatPartitionColumn(change,jd,list,tableName);
+                    hiveAlterTable.formatClusterColumn(change,jd,list,tableName);
 
                 } catch (Exception ex) {
                     log.warn("Error comparing formats", ex);
@@ -113,240 +119,8 @@ public class HiveHqlGenerator {
 
 
 
-    void formatRegularColumn(FormatChange change, JobDefinition jd, List<String> list, String tableName) throws org.finra.herd.sdk.invoker.ApiException
-    {
-        if (change.hasColumnChanges()) {
-            boolean cascade = cascade(jd);
-            String cascadeStr = "";
-            if (cascade) {
-                cascadeStr = "CASCADE";
-            }
-            if (!jd.getObjectDefinition().getFileType().equalsIgnoreCase("ORC")) {
-                String sql = dataMgmtSvc.getTableSchema(jd, true);
 
 
-                if (cascade) {
-                    sql = sql.substring(0, sql.lastIndexOf(";")) + " CASCADE;";
-                }
-                list.add(sql);
-            } else {
-                for (Pair<ColumnDef, ColumnDef> pair : change.getNameChanges()) {
-                    ColumnDef existing = pair.getFirst();
-                    ColumnDef newColum = pair.getSecond();
-
-                    list.add(String.format("Alter table %s change %s %s %s %s;", tableName,
-                            existing.getName(), newColum.getName(), newColum.getType(), cascadeStr));
-
-
-
-                }
-
-                for (Pair<ColumnDef, ColumnDef> pair : change.getTypeChanges()) {
-                    ColumnDef existing = pair.getFirst();
-                    ColumnDef newColum = pair.getSecond();
-                    list.add(String.format("Alter table %s change %s %s %s %s;", tableName,
-                            existing.getName(), newColum.getName(), newColum.getType(), cascadeStr));
-                }
-
-                if (!change.getNewColumns().isEmpty()) {
-                    StringBuffer sb = new StringBuffer();
-                    for (ColumnDef c : change.getNewColumns()) {
-                        sb.append(String.format("%s %s,", c.getName(), c.getType()));
-                    }
-                    sb.deleteCharAt(sb.length() - 1);
-                    list.add(String.format("Alter table %s add columns (%s) %s;", tableName, sb.toString(),
-                            cascadeStr));
-                }
-
-
-                log.info("the formatRegularColumn list is :{}",list);
-            }
-        }
-
-
-    }
-
-    void formatClusterColumn(FormatChange change,JobDefinition jd,List<String> list,String tableName) throws org.finra.herd.sdk.invoker.ApiException {
-
-        if(change.isClusteredSortedChange())
-        {
-
-                    log.info("Clustered Change :Alter table {}  {}", tableName, change.getClusteredDef().getClusterSql());
-                    list.add(String.format("Alter table %s  %s ;", tableName,
-                            change.getClusteredDef().getClusterSql()));
-
-
-        }
-
-    }
-
-    void formatPartitionColumn(FormatChange change, JobDefinition jd, List<String> list, String tableName) throws org.finra.herd.sdk.invoker.ApiException{
-
-        if(change.hasPartitionColumnChanges())
-        {
-
-
-                for (Pair<ColumnDef, ColumnDef> pair : change.getPartitionColTypeChanges()) {
-                    ColumnDef existing = pair.getFirst();
-                    ColumnDef newColum = pair.getSecond();
-                    list.add(String.format("Alter table %s partition column ( %s %s);", tableName,
-                             newColum.getName(), newColum.getType()));
-                }
-
-
-                log.info("the formatPartitionColumn list is :{}",list);
-
-        }
-
-
-    }
-
-
-
-
-    @VisibleForTesting
-    FormatChange detectSchemaChange(JobDefinition jd,
-                                    HiveTableSchema existingHiveTableSchema,
-                                    BusinessObjectFormat format, String ddl) throws ApiException {
-
-
-        HiveTableSchema newSchema = HiveClientImpl.getHiveTableSchema(ddl);
-        List<ColumnDef> existingColumns = existingHiveTableSchema.getColumns();
-        List<ColumnDef> existingPartitionColumns = existingHiveTableSchema.getPartitionColumns();
-        List<ColumnDef> newColumns = newSchema.getColumns();
-        List<ColumnDef> newPartitionColumns = newSchema.getPartitionColumns();
-        ClusteredDef existingClusteredDef = existingHiveTableSchema.getClusteredDef();
-        ClusteredDef newClusterDef = newSchema.getClusteredDef();
-
-
-        log.info("Existing Partition columns = " + existingPartitionColumns.size() + ", ddl from Herd  Partitioncolumns = " + newPartitionColumns.size());
-
-        log.info("Existing columns = " + existingColumns.size() + ", ddl from Herd has columns = " + newColumns.size());
-
-
-        FormatChange formatChange = FormatChange.builder().build();
-
-        detectandSetRegularColumnChanges(existingColumns,newColumns,formatChange);
-        detectandSetPartitionColumnChanges(existingPartitionColumns,newPartitionColumns,formatChange);
-
-        if(detectClusterSortedColChanges(existingClusteredDef,newClusterDef))
-        {
-            formatChange.setClusteredSortedChange(true);
-            formatChange.setClusteredDef(newClusterDef);
-        }
-
-
-        if (formatChange.hasChange() ) {
-            notificationSender.sendFormatChangeEmail(formatChange, format.getBusinessObjectFormatVersion(), jd,
-                    existingHiveTableSchema, newSchema);
-        }
-
-        return formatChange;
-    }
-
-    boolean detectClusterSortedColChanges(ClusteredDef existing,ClusteredDef recent){
-
-        boolean isClusterSortedChg = false;
-        List<ColumnDef> existingColumns = existing.getClusteredSortedColDefs();
-        List<ColumnDef> newColumns = recent.getClusteredSortedColDefs();
-        log.info("existing Cluster by Sorted by Columns:{}",existingColumns);
-        log.info("recent Cluster by Sorted by Columns:{} ",newColumns);
-
-      if ((existingColumns !=null && !existingColumns.isEmpty()) && (newColumns !=null && !newColumns.isEmpty()) )  {
-          int minColumns = Math.min(existingColumns.size(),newColumns.size());
-
-          for (int i = 0; i < minColumns; i++) {
-              ColumnDef old = existingColumns.get(i);
-              ColumnDef newColum = newColumns.get(i);
-
-              if (!newColum.getName().equalsIgnoreCase(old.getName()) )
-              {
-                  isClusterSortedChg = true;
-              }
-              }
-
-      }else if ((existingColumns == null || existingColumns.isEmpty()) && (newColumns !=null && !newColumns.isEmpty())  )
-      {
-          isClusterSortedChg = true;
-
-      }
-
-      log.info("is there a Cluster or Sorted Column Change?:{}",isClusterSortedChg);
-      return isClusterSortedChg;
-    }
-
-    @VisibleForTesting
-    void detectandSetRegularColumnChanges(List<ColumnDef> existingColumns, List<ColumnDef> newColumns,FormatChange formatChange)
-    {
-        List<Pair<ColumnDef, ColumnDef>> nameChanges = Lists.newArrayList();
-        List<Pair<ColumnDef, ColumnDef>> typeChanges = Lists.newArrayList();
-        List<ColumnDef> addedColumns = Lists.newArrayList();
-
-
-
-        int minColumns = Math.min(existingColumns.size(), newColumns.size());
-
-        //@TODO - Refactor and Move all of these logic out of HqlGenerator when we implement Format Change api using sns
-            /*
-              Regular Column
-             */
-        for (int i = 0; i < minColumns; i++) {
-            ColumnDef existing = existingColumns.get(i);
-            ColumnDef newColum = newColumns.get(i);
-
-            if (!newColum.getName().equalsIgnoreCase(existing.getName())) {
-                nameChanges.add(new Pair<>(existing, newColum));
-            } else if (!newColum.isSameType(existing)) {
-                typeChanges.add(new Pair<>(existing, newColum));
-            }
-
-        }
-
-        if (newColumns.size() > existingColumns.size()) {
-            for (int i = existingColumns.size(); i < newColumns.size(); i++) {
-                ColumnDef newColum = newColumns.get(i);
-                addedColumns.add(newColum);
-            }
-        }
-
-        log.info("Regular Column Changes nameChanges :{}, typeChanges:{},addedColumns :{}",nameChanges,typeChanges,addedColumns);
-
-        formatChange.setNameChanges(nameChanges);
-        formatChange.setTypeChanges(typeChanges);
-        formatChange.setNewColumns(addedColumns);
-
-
-    }
-
-    @VisibleForTesting
-    void detectandSetPartitionColumnChanges(List<ColumnDef> existingPartitionColumns, List<ColumnDef> newPartitionColumns,FormatChange formatChange)
-    {
-        List<Pair<ColumnDef, ColumnDef>> partitionColTypeChanges = Lists.newArrayList();
-        List<Pair<ColumnDef, ColumnDef>> partitionColNameChanges = Lists.newArrayList();
-
-
-        int minColumns = Math.min(existingPartitionColumns.size(), newPartitionColumns.size());
-
-
-        for (int i = 0; i < minColumns; i++) {
-            ColumnDef existing = existingPartitionColumns.get(i);
-            ColumnDef newColum = newPartitionColumns.get(i);
-
-            if (!newColum.getName().equalsIgnoreCase(existing.getName())) {
-                partitionColNameChanges.add(new Pair<>(existing, newColum));
-                log.error("Hive does not support partition Column Name changes:{}",partitionColNameChanges);
-
-            } else if (!newColum.isSameType(existing)) {
-                partitionColTypeChanges.add(new Pair<>(existing, newColum));
-            }
-
-        }
-
-        formatChange.setPartitionColNameChanges(partitionColNameChanges);
-        formatChange.setPartitionColTypeChanges(partitionColTypeChanges);
-
-
-    }
 
 
 
@@ -452,9 +226,6 @@ public class HiveHqlGenerator {
                 .build();
     }
 
-    protected boolean cascade(JobDefinition jd) {
-        return true;
-    }
 
     /**
      * Method to create the HQL file for adding partitions
